Move func (b *BlockRange) MarshalJSON() to misc.go
